dry_run for loaders

+++ py/src/costume/etl/_loader.py
@@ -10,9 +10,10 @@ from services import Services
 
 
 class _Loader(_Main):
-    def __init__(self, institution_id, clean=False, **kwds):
+    def __init__(self, institution_id, clean=False, dry_run=False, **kwds):
         _Main.__init__(self, **kwds)
         self.__clean = clean
+        self.__dry_run = dry_run
         self.__institution_id = institution_id
         self.__properties = CostumeProperties.load()
         self.__services = Services(properties=self.__properties)
@@ -20,8 +21,12 @@ class _Loader(_Main):
     @classmethod
     def _add_arguments(cls, argument_parser):
         argument_parser.add_argument('--clean', action='store_true')
+        argument_parser.add_argument('--dry-run', action='store_true')
 
     def _clean(self):
+        if self.__dry_run:
+            self._logger.info("dry run, not deleting insittution %s", self._institution_id)
+            return
         try:
             self._services.institution_command_service.delete_institution_by_id(self._institution_id)
         except NoSuchInstitutionException:
@@ -55,7 +60,7 @@ class _Loader(_Main):
     def _run(self):
         if self.__clean:
             self._clean()
-        self._load()
+        self._load(dry_run=self.__dry_run)
 
     @property
     def _services(self):
